+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # 0.2.3 (feb 25th 2025)
 
+- Improve error handling of PWD compare() method
 - Fix declaration file export for Typescript
 
 # 0.2.2 (feb 24th 2025)
 
-- Improve JWT verify() method error handling
+- Improve error handling of JWT verify() method
 
 # 0.2.1 (feb 23th 2025)